feat: update pokemon weight data type

+++ app/src/main/java/com/fakhry/pokedex/domain/model/Pokemon.kt
@@ -3,7 +3,7 @@ package com.fakhry.pokedex.domain.model
 data class Pokemon(
     val id: Int,
     val name: String,
-    val weight: Int = 0,
+    val weight: Double = 0.0,
     val frontImage: String = "",
     val pictures: List<String> = listOf(),
     val types: List<PokemonType> = listOf(),
